examples: fix g_object_set() value type.
Make sure we cast the length value as a gint64 to the vararg g_object_set() just
incase sizeof(gsize) != sizeof(gint64).

+++ tests/examples/app/appsrc-seekable.c
@@ -128,7 +128,7 @@ found_source (GObject * object, GObject * orig, GParamSpec * pspec, App * app)
   /* we can set the length in appsrc. This allows some elements to estimate the
    * total duration of the stream. It's a good idea to set the property when you
    * can but it's not required. */
-  g_object_set (app->appsrc, "size", app->length, NULL);
+  g_object_set (app->appsrc, "size", (gint64) app->length, NULL);
   /* we are seekable in push mode, this means that the element usually pushes
    * out buffers of an undefined size and that seeks happen only occasionally
    * and only by request of the user. */
